Implement 3D metric with topography
This commit implements the 3D metric, including
the terrain-following coordinate system.

Because of the underlying code structure, we define the
geometry object (cubed_sphere) and the corresponding metric
before we have the topography in hand.  This in turn
requires that we split the computation into two phases, with
recomputation after defining the topography.

For the geometry object, this recomputation is effected by
the apply_topography() member, which redefines the x3
coordinate (as physical height, in meters) and recomputes
the other coordinate arrays.

The geometry object also now contains the consolidated
coordinate arrays of coordVec_gnom, _cart, _latlon, and _num
(numerical/equiangular), along with _itf_i/j/k varieties.

For the metric object (Metric_3d_topo), the metric
initialization is only a stub, and computation is deferred
until a call to Metric_3d_topo.build_metric().  This avoids
evaluating more-expensive metric computations that would
only be overwritten.

The metric object now also contains the consolidated arrays
of H_cov (covariant) and H_contra (contravariant), with
shapes of [3,3,nk,nj,ni], so that indices can be raised or
lowered via an ordinary linear algebra expression.

Additionally, the matrix object (DFR_Operators), now
contains built-in methods to take derivatives along each of
the three dimensions when given element-boundary values.