Add boundary exchange for topography gradient in metric
This commit adds the previously-neglected exchange of
gradients at element (parallel process) boundaries in the
metric calculation.  Previously, for parallel boundaries the
metric was based on one-sided derivatives of topography
(height), which could cause an inconsistency in flux
calculations (which use the metric at element boundaries)
and the Christoffel symbols (which use the derivative of the
metric).