Add exponential filter post-timestep
This commit adds the option to apply an exponential
mode-based filter after each timestep.  Doing so is
necessary for stability in the DCMIP21 case (with RK3
timestepping).  This filter follows the form of equation
(5.16) in Warburtion (section 5.3), where in each dimension
it has the form:

F̃(k) = F exp(-α ((k-kc)/(1-kc))^p)

where α is the filter strength, kc is the cutoff frequency
below which no filtering is applied, and p is the filter
order.  These parameters correspond to `expfilter_strength`,
`expfilter_cutoff`, and `expfilter_order` in the
configuration file.  Modes are normalized between 0 and 1,
inclusive.

This filter is applied isotopically via the tensor product
structure.  To preserve mass and momentum, each variable is
multiplied by √g before filtering and 1/√g after.  This is
also more compatible with the differentiation inside the
Euler equations, where the flux term also involves √g
factors.