Metric: don't perform boundary exchange for nproc=1
It's occasionally been convenient to debug the Euler
equations by buidling a single-panel domain in a Jupyter
notebook.  This debugging re-uses the matrix and metric
objects, shimming the geometry and ptopo objects.

This only works, however, if the boundary-exchange of
topography derivatives is disabled.  In the single-processor
case, there are no boundary processors, and calls to the MPI
object will fail.  This commit introduces special-case code
inside the metric object to bypass MPI calls when nproc=1.