rhs_euler, use log(pressure) for w-momentum equation
This commit modifies the vertical momentum equation to
evaluate the derivatives of pressure as pres*D(log(pres)).
This form is mathematically equivalent to ordinary
differentiation, but it is numerically better-behaved with
height changes, since log(pressure) is close to linear in
height whereas pressure itself has an exponential
dependence.

This commit also modifies the extrapolation of ρ and ρθ to
extrapolate the logarithm of both, finding the
element-boundary value after exponentiation.  This change
reduces spurious diffusion of mass and entropy (potential
temperature) in the atmosphere at rest with well-resolved
topography.