Add option for shallow atmosphere approximation
This commit adds support for the shallow atmosphere
approximation for the Euler equations on the sphere.  In
this approximation, the metric terms that depend on the
radial coordinate r (not its derivatives) are replaced by
geom.earth_radius.

This approximation matches that currently used by GEM, and
it is specified for the DCMIP test cases.  In particular,
test case 2-1 is not in balance in a deep atmosphere,
despite the physical parameters of a reduced-size planet
implying a very deep atmosphere.